JCR-2085 TCK
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@779621 13f79535-47bb-0310-9956-ffa450edef68

+++ jackrabbit-jcr-tests/src/main/java/org/apache/jackrabbit/test/api/BinaryPropertyTest.java
@@ -438,7 +438,7 @@ public class BinaryPropertyTest extends AbstractPropertyTest {
             int k = 0;
             for (int b; (b = in.read()) != -1; k++) {
                 assertEquals("Value.getStream().read() and Value.getBinary().read() " +
-                        "return different values.", b, content[k]);
+                        "return different values.", (byte) b, content[k]);
             }
             assertEquals("unexpected content length", k, content.length);
         } finally {
